+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = [
     {name = "Yaroslav",email = "zabaraaroslav@gmail.com"}
 ]
 readme = "README.md"
-requires-python = ">=3.13,<4"
+requires-python = ">=3.12,<4"
 dependencies = [
     "rsa (>=4.9,<5.0)"
 ]
@@ -15,6 +15,9 @@ dependencies = [
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.poetry]
+package-mode = false
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.5"
